+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/TestActivity.kt
@@ -1,0 +1,5 @@
+package com.rakuten.tech.mobile.miniapp
+
+import android.app.Activity
+
+class TestActivity : Activity()